% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmol.R
\docType{class}
\name{Chemmol-class}
\alias{Chemmol-class}
\alias{Chemmol}
\title{Chemmol Class}
\description{
An S4 class to represent chemical molecules using SDF and SMILES formats.
}
\section{Slots}{

\describe{
\item{\code{sdf}}{An object of class `SDFset` containing the SDF representation.}

\item{\code{smiles}}{An object of class `SMIset` containing the SMILES representation.}
}}

\examples{
sdf <- read.SDFset("path_to_sdf_file")
smiles <- c("CCO", "CCC")
chem <- Chemmol(sdf = sdf, smiles = smiles)
}
