% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmol.R
\name{dist,Chemmol-method}
\alias{dist,Chemmol-method}
\title{Calculate Distance Between Chemmol Objects}
\usage{
\S4method{dist}{Chemmol}(x, method = "FP2")
}
\arguments{
\item{x}{An object of class `Chemmol`.}

\item{method}{The fingerprint method to use. Default is "FP2".}
}
\value{
A distance matrix.
}
\description{
Compute a distance matrix for objects of class `Chemmol` using specified fingerprint method.
}
\examples{
chem <- Chemmol(sdf = sdf, smiles = smiles)
dist_matrix <- dist(chem, method = "FP2")
}
