% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmol.R
\name{generate_chem_glyphs}
\alias{generate_chem_glyphs}
\title{Generate Chemical Glyphs}
\usage{
generate_chem_glyphs(
  smiles,
  outdir = tempdir(),
  bond_col = "black",
  show_atoms = TRUE,
  resolution = 200,
  node_background_color = rgb(1, 1, 1, 1),
  font_size = 5,
  ...
)
}
\arguments{
\item{smiles}{A character vector of SMILES strings.}

\item{outdir}{Directory where the output PNG files will be saved. Default is a temporary directory.}

\item{bond_col}{Color of the bonds in the chemical structure. Default is "black".}

\item{show_atoms}{Logical value indicating whether to show atom labels. Default is TRUE.}

\item{resolution}{Resolution of the output images. Default is 200.}

\item{...}{Additional arguments passed to `ChemmineR::plot`.}
}
\value{
A character vector of file paths to the generated PNG images.
}
\description{
This function generates PNG images of chemical structures from SMILES strings using the `ChemmineR` package.
}
\examples{
smiles <- c("CCO", "CCC")
generate_chem_glyphs(smiles)
}
