% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemmap.R
\name{geom_chemmap}
\alias{geom_chemmap}
\title{Title}
\usage{
geom_chemmap(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
