% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmol.R
\name{geom_mol}
\alias{geom_mol}
\title{GeomMol: A ggplot2 Geom for Plotting Chemical Structures}
\usage{
geom_mol(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by `aes()`.}

\item{data}{The data to be displayed in this layer.}

\item{stat}{The statistical transformation to use on the data for this layer. Default is "identity".}

\item{position}{Position adjustment. Default is "identity".}

\item{na.rm}{If `FALSE`, removes missing values with a warning. If `TRUE`, silently removes missing values. Default is FALSE.}

\item{show.legend}{Logical. Should this layer be included in the legends? Default is `NA`.}

\item{inherit.aes}{If `FALSE`, overrides the default aesthetics. Default is `TRUE`.}

\item{...}{Other arguments passed on to `layer()`.}
}
\value{
A `ggplot2` layer.
}
\description{
This `ggplot2` geom adds chemical structure images to a plot based on SMILES strings.
}
\examples{
library(ggplot2)
smiles <- c("CCO", "CCC")
df <- data.frame(x = 1:2, y = 1:2, smiles = smiles)
ggplot(df, aes(x, y, smiles = smiles)) +
  geom_mol()
}
