% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmol.R
\docType{data}
\name{GeomMol}
\alias{GeomMol}
\title{GeomMol ggproto Object}
\format{
An object of class \code{GeomMol} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomMol
}
\description{
The ggproto object defining the GeomMol for ggplot2.
}
\keyword{datasets}
