% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmol.R
\name{initialize,Chemmol-method}
\alias{initialize,Chemmol-method}
\title{Initialize Chemmol Object}
\usage{
\S4method{initialize}{Chemmol}(.Object, sdf = NULL, smiles = NULL, ...)
}
\arguments{
\item{.Object}{The `Chemmol` object to initialize.}

\item{sdf}{An optional `SDFset` object.}

\item{smiles}{An optional character vector of SMILES strings.}

\item{...}{Additional arguments passed to `callNextMethod()`.}
}
\description{
Initialize a new object of class `Chemmol`.
}
